+++ backend/src/main/java/diyor/adawev/backend/dto/MaterialItemResponse.java
@@ -1,7 +1,6 @@
 package diyor.adawev.backend.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -10,7 +9,7 @@ import java.math.BigDecimal;
 public class MaterialItemResponse {
     private Long id;
     private String materialName;      // Odam o'qiydigan nom (edit qilish mumkin)
-    private BigDecimal quantity;       // Miqdor (edit qilish mumkin)
+    private Float quantity;       // Miqdor (edit qilish mumkin)
     private String unit;
     private String originalName;       // Database'dagi original nom (reference uchun)
     private String type;               // Material type (reference uchun)
